+++ elements/pl_figure/pl_figure.py
@@ -7,7 +7,6 @@ import os
 def prepare(element_html, element_index, data):
     element = lxml.html.fragment_fromstring(element_html)
     pl.check_attribs(element, required_attribs=['file_name'], optional_attribs=['width', 'type', 'directory'])
-    return data
 
 
 def render(element_html, element_index, data):