+++ elements/pl_file_download/pl_file_download.py
@@ -6,7 +6,6 @@ import os
 def prepare(element_html, element_index, data):
     element = lxml.html.fragment_fromstring(element_html)
     pl.check_attribs(element, required_attribs=['file_name'], optional_attribs=['type', 'directory', 'label'])
-    return data
 
 
 def render(element_html, element_index, data):