+++ elements/pl_variable_score/pl_variable_score.py
@@ -7,11 +7,10 @@ use_pl_variable_score = False
 
 def prepare(element_html, element_index, data):
     if not use_pl_variable_score:
-        return data
+        return
 
     element = lxml.html.fragment_fromstring(element_html)
     pl.check_attribs(element, required_attribs=['answers_name'], optional_attribs=[])
-    return data
 
 
 def render(element_html, element_index, data):