+++ exampleCourse/questions/addNumbers/server.py
@@ -1,13 +1,17 @@
-import random
+import random, copy
 
 def generate(data):
 
+    # Sample two random integers between 5 and 10 (inclusive)
     a = random.randint(5, 10)
     b = random.randint(5, 10)
-    data["params"]["a"] = a
-    data["params"]["b"] = b
 
+    # Put these two integers into data['params']
+    data['params']['a'] = a
+    data['params']['b'] = b
+
+    # Compute the sum of these two integers
     c = a + b
-    data["correct_answers"]["c"] = c
 
-    return data
+    # Put the sum into data['correct_answers']
+    data['correct_answers']['c'] = c